v0.8 Sprint 1 (QA-014): decompose OutboxOrchestrator — close GodClass + TooManyMethods + CyclomaticComplexity + AvoidCatchingGenericException (#128)
Sprint 1 GodClass closure track continues. OutboxOrchestrator (461 LOC,
4 class-level PMD suppressions: TooManyMethods + AvoidCatching
GenericException + CyclomaticComplexity + GodClass) splits along two
cohesive seams:

  - State machine: VarHandle CAS state primitive + JFR transition
    emission + six ordinal/name pair encoding
  - Batch flush: broker.publish + store.markDelivered + per-event
    retry with linear backoff + DLQ routing + JFR batch/DLQ events

Three files now:

  OutboxStateMachine (133 LOC, pkg-private)
    * STATE_* ordinal constants (IDLE/POLLING/FLUSHING/WAITING/RETRYING/STOPPED)
    * VarHandle STATE_VH + volatile int state field
    * transitionTo(next, polledCount) — CAS guarded by runningCheck BooleanSupplier
    * forceTransitionToStopped() — getAndSet for terminal transition from stop()
    * currentStateName() / currentOrdinal() / isStopped() accessors
    * Internal JFR emit on every successful transition

  OutboxBatchFlusher (171 LOC, pkg-private)
    * flush(batch) — broker.publish, markDelivered, partial-failure routing
    * handlePartialFailure — per-event retry, DLQ on exhaustion
    * retryEntry — single-event retry with linear backoff
    * Internal closeEntryPayloads + JFR batch/DLQ emission

  OutboxOrchestrator (276 LOC, public — refactored)
    * Public surface UNCHANGED: builder(), start(), stop(), close(),
      currentState() string + Builder fluent setters all preserved
    * ownerLoop (StructuredTaskScope owner VT — Java 26 owner-thread rule)
    * runLoop + executeTick — drives the state machine through the flusher
    * Builder static inner class — verbatim

Suppression delta:

  Closed (4 class-level on OutboxOrchestrator):
    - PMD.TooManyMethods
    - PMD.AvoidCatchingGenericException  (one residual per-method on executeTick)
    - PMD.CyclomaticComplexity
    - PMD.GodClass

  New suppressions on OutboxBatchFlusher (with inline rationale):
    - PMD.AvoidCatchingGenericException — broker.publish() throws RuntimeException by contract
    - PMD.CloseResource — payload close is routed through closeEntryPayloads

  New per-method suppressions on OutboxOrchestrator:
    - constructor: PMD.LawOfDemeter (canonical Builder pattern)
    - executeTick: PMD.AvoidCatchingGenericException + PMD.LawOfDemeter

  Net: -4 class-level suppressions, +2 class-level on new helper, +2
       method-level — but the new ones are narrowly scoped with
       documented rationale, vs the original 4 covering 461 LOC.

Public API surface unchanged. Tests use only currentState() string +
builder() + lifecycle methods (verified: OutboxOrchestratorTest,
OutboxOrchestratorIntegrationTest, CommunityOutboxOrchestratorTckTest,
CommunityOutboxGuaranteeTckTest, CommunityEventEngineOutboxIntegration
Test).

Delta:
  OutboxOrchestrator 461 → 276 LOC (-185 / -40%)
  + OutboxStateMachine 133 LOC
  + OutboxBatchFlusher 171 LOC
  Class-level suppressions on orchestrator: 4 → 0 (-4, full closure)

Test plan:
  - OutboxOrchestratorTest 6/6 green (state lifecycle + currentState semantics)
  - OutboxOrchestratorIntegrationTest 1/1 green (start → poll → stop sequence)
  - CommunityOutboxOrchestratorTckTest 4/4 green (community TCK binding)
  - CommunityOutboxGuaranteeTckTest 2/2 green (at-least-once + replay idempotency)
  - Full reactor verify SUCCESS (SPI + TCK + Core + Community Testkit +
    Community + Build Config + Parent + Root)
  - PMD 0 violations on core module
  - Checkstyle 0 violations on core module
  - Local JDK 26+35 G1 crash worked around with -XX:+UseSerialGC
    (pre-existing memory note; unrelated to this refactor)
